network-score: change methods to suit the need of API spec
with more discussions, it was aggreed that it is entity writing to
chain's task to build the payload.

In some cases, as in the example script, the validation of source
is done by the application itself (example, the API for ONDC score
may be done by Beckn signature check methods).

Signed-off-by: Amar Tumballi <amar@dhiway.com>